Setting up

EasyML.Common.change — Method

change(global_options::GlobalOptions)

Allows to change global_options in a GUI.

EasyML.Common.load_options — Function

load_options()

Loads options from your previous run which are located in options.bson. Uses present working directory. It is run automatically after loading the package.

EasyML.Common.save_options — Function

save_options()

Saves options to options.bson. Uses present working directory. It is run automatically after changing options in a GUI window.

EasyML.Common.save_model — Function

save_model(url::String)

Saves a model to a specified URL. The URL can be absolute or relative. Use '.model' extension.

save_model()

Opens a file dialog where you can select where to save a model and how it should be called.

EasyML.Common.load_model — Function

load_model(url::String)

Loads a model from a specified URL. The URL can be absolute or relative.

load_model()

Opens a file dialog where you can select a model to be loaded and loads it.

Design

EasyML.Classes.change_classes — Method

change_classes()

Opens a GUI for addition or changing of classes.

EasyML.Design.design_model — Method

design_model()

Opens a GUI for creation of a model.

Training

EasyML.Common.change — Method

change(data_preparation_options::DataPreparationOptions)

Allows to change data_preparation_options in a GUI.

EasyML.Common.change — Method

change(training_options::TrainingOptions)

Allows to change training_options in a GUI.

EasyML.get_urls_training — Function

get_urls_training(url_inputs::String,url_labels::String)

Gets URLs to all files present in both folders (or a folder and a file) specified by url_inputs and url_labels for training. URLs are automatically saved to EasyML.training_data.

get_urls_training(url_inputs::String)

Used for classification. Gets URLs to all files present in folders located at a folder specified by url_inputs for training. Folders should have names identical to the name of classes. URLs are automatically saved to EasyML.training_data.

get_urls_training()

Opens a folder/file dialog or dialogs to choose folders or folder and a file containing inputs and labels. URLs are automatically saved to EasyML.training_data.

EasyML.get_urls_testing — Function

get_urls_testing(url_inputs::String,url_labels::String)

Gets URLs to all files present in both folders (or a folder and a file) specified by url_inputs and url_labels for testing. URLs are automatically saved to EasyML.testing_data.

get_urls_testing(url_inputs::String)

Used for classification. Gets URLs to all files present in folders located at a folder specified by url_inputs for testing. Folders should have names identical to the name of classes. URLs are automatically saved to EasyML.testing_data.

get_urls_testing()

If testing data preparation in modify(training_options) is set to auto, then a percentage of training data also specified there is reserved for testing. If testing data preparation is set to manual, then it opens a folder/file dialog or dialogs to choose folders or a folder and a file containing inputs and labels. URLs are automatically saved to EasyML.testing_data.

EasyML.prepare_training_data — Function

prepare_training_data()

Prepares images and corresponding labels for training using URLs loaded previously using get_urls_training. Saves data to EasyML.training_data.

EasyML.prepare_testing_data — Function

prepare_testing_data()

Prepares images and corresponding labels for testing using URLs loaded previously using get_urls_testing. Saves data to EasyML.testing_data.

EasyML.Training.train — Function

train()

Opens a GUI where training progress can be observed. Training parameters such as a number of epochs, learning rate and a number of tests per epoch can be changed during training.

EasyML.Training.remove_training_data — Function

remove_training_data()

Removes all training data except for result.

EasyML.Training.remove_testing_data — Function

remove_testing_data()

Removes all testing data.

EasyML.Training.remove_training_results — Function

remove_training_results()

Removes training results.

Validation

EasyML.Common.change — Method

change(validation_options::ValidationOptions)

Allows to change validation_options in a GUI.

EasyML.Validation.get_urls_validation — Function

get_urls_validation(url_inputs::String,url_labels::String)

Gets URLs to all files present in both folders (or a folder and a file) specified by url_inputs and url_labels for validation. URLs are automatically saved to EasyMLValidation.validation_data.

get_urls_validation(url_inputs::String)

Gets URLs to all files present in a folders specified by url_inputs for validation. URLs are automatically saved to EasyMLValidation.validation_data.

get_urls_validation()

Opens a folder/file dialog or dialogs to choose folders or folder and a file containing inputs and labels. Folder/file dialog for labels can be skipped if there are no labels available. URLs are automatically saved to EasyMLValidation.validation_data.

EasyML.Validation.validate — Function

validate()

Opens a GUI where validation progress and results can be observed.

EasyML.Validation.remove_validation_data — Function

remove_validation_data()

Removes all validation data except for result.

EasyML.Validation.remove_validation_results — Function

remove_validation_results()

Removes validation results.

Application

EasyML.Common.change — Method

change(application_options::EasyML.ApplicationOptions)

Allows to change application_options in a GUI.

EasyML.Application.get_urls_application — Function

get_urls_application(url_inputs::String)

Gets URLs to all files present in a folders specified by url_inputs for application. URLs are automatically saved to EasyML.application_data.

get_urls_application()

Opens a folder dialog to choose a folder containing files to which a model should be applied. URLs are automatically saved to EasyML.application_data.

EasyML.Application.apply — Function

apply()

Starts application of a model.

EasyML.Application.remove_application_data — Function

remove_application_data()

Removes all application data.